CRASH-AUDIT: диагностика нативных падений воркера и закрытие критичных находок
Пакет правок по разбору SIGABRT на Apple Silicon; отчёт — reports/crash-audit-report.md.

Диагностика падений (P1):
- воркер пишет Python-стек нативного падения в logs/worker_faulthandler.log: без
  этого SIGABRT не оставлял ни строки, и причину восстанавливали по форме кадров
  в .ips;
- при старте воркер логирует pid, argv, executable, cwd и sys.path — в дампе
  виден только стек, а по этим полям ясно, кто именно упал;
- sys.excepthook и threading.excepthook в воркере и в бэкенде: исключения вне
  handle_request и в потоках движка раньше не попадали в лог вовсе;
- заглушка tests/fake_worker_engine.py печатает метку FAKE-WORKER-CRASH перед
  os.abort(), а родитель переносит stderr воркера в свой лог, поэтому тестовое
  падение опознаётся по одной строке.

Правила и инструменты (P2/P3):
- docs/testing.md: тесты изоляции и замеры памяти гоняются из Terminal.app, плюс
  регламент «как отличить тестовый SIGABRT от продуктового за 1 минуту»;
- docs/architecture.md: sklearn/onnxruntime/df/libdf не живут в одном процессе с
  torch (дубль libomp даёт настоящий SIGABRT), инвариант закреплён тестом;
- docs/operations.md: отчёты macOS о падениях — окно наблюдения, а не архив;
- tests/test_backend_invariants.py: запрет os.abort/os._exit/faulthandler.disable
  и set_start_method("fork") в backend/, запрет импорта тяжёлых нативных пакетов;
- tools/retire_test_crashes.py: разбор Python-*.ips после прогона, подтверждение
  метки по журналу, перенос в Retired/test-crashes/ или удаление.

Закрытые критичные находки: атомарность миграций и voices.json, проверка пути
референса внутри voices/, защита models/ при удалении F5, незакрытая задача при
срабатывании watchdog, валидация мусорных полей в ответах LLM, HeavyResourceBusy
в очереди, ключ кеша разборов, pin прямо объявленных зависимостей и один
минорный ряд torch/torchaudio.

Уборка: отчёт аудита переехал в reports/ (индекс — в docs/history.md), ТЗ-файлы
расследования из корня удалены, в истории поправлена опечатка creash_ -> crash_.

Проверки: pytest — полный прогон зелёный (exit 0), ruff чисто; живое приложение:
device=mps, синтез дал tts.mps.synthesis delta_mb=-616.4, выгрузка движка
остановила воркер, новых Python-*.ips за сессию нет.

+++ backend/engines/base.py
@@ -156,7 +156,7 @@ class EngineParam:
         except (TypeError, ValueError):
             number = float(self.default)
         number = min(max(number, self.minimum), self.maximum)
-        return int(round(number)) if self.integer else round(number, 4)
+        return round(number) if self.integer else round(number, 4)
 
     def to_dict(self) -> dict:
         return {
